Simplify the address type setter
This moves the functionality that sets the address type
based on payment type into a composable.

That means we can use the same form step controllers in
CTRL and VAR and no longer need 2 new ones.

Tickets:
- https://phabricator.wikimedia.org/T353268
- https://phabricator.wikimedia.org/T352973
