--- src/components/DonationForm/StepControllers/SubmittableMainDonationFormAnonymous.ts
@@ -1,29 +1,0 @@
-import { FormModel } from '@src/utils/FormModel/FormModel';
-import { Intervals } from '@src/utils/FormItemsBuilder/fields/Intervals';
-import { PaymentMethods } from '@src/utils/FormItemsBuilder/fields/PaymentMethods';
-import { StepController } from '@src/components/DonationForm/StepController';
-import { StepAction } from '@src/components/DonationForm/StepNavigation';
-import { BannerSubmitEvent } from '@src/tracking/events/BannerSubmitEvent';
-import { AddressTypes } from '@src/utils/FormItemsBuilder/fields/AddressTypes';
-
-export function createSubmittableMainDonationFormAnonymous( formModel: FormModel, stepNameOfUpgradeToYearly: string ): StepController {
-	return {
-		async submit( navigation: StepAction ): Promise<void> {
-			if ( formModel.interval.value !== Intervals.ONCE.value || formModel.paymentMethod.value === PaymentMethods.SOFORT.value ) {
-				if ( !paymentMethodRequiresAddress( formModel ) ) {
-					formModel.addressType.value = AddressTypes.ANONYMOUS.value;
-				}
-				await navigation.submit( new BannerSubmitEvent( 'MainDonationForm' ) );
-				return;
-			}
-			await navigation.goToStep( stepNameOfUpgradeToYearly );
-		},
-		async previous(): Promise<void> {
-			return Promise.reject( 'we can\'t go to previous! This should never happen' );
-		}
-	};
-}
-
-function paymentMethodRequiresAddress( formModel: FormModel ): boolean {
-	return formModel.paymentMethod.value === PaymentMethods.DIRECT_DEBIT.value;
-}
